Created a multipass shader with uniforms
In order to setup a velocity buffer I need to be able to computer
changes per vertex each render call. However, in order to do this I need
to get the previous view-projection matrix that is stored by a script.
The next problem is that since this is no longer a post-process effect
that is rendered from a texture I need to apply this to objects. I still
want to keep the shaders that Unity has made so I created a shader that
does two passes, the first pass is with the unity shader and the second
is with my own vertex and fragment shader on top of that.