Created the velocity buffer and renders it to screen
Created the basic velocity buffer and render it to screen. The next step
is to render it to a texture and then read from that texture when
calculating the blurring. I have to find a good sampling algorithm for
the blurring as well. I need to look into deltaTime from rendering calls
as well since the velocity buffer produces various results depending on
the frame rate.