Access velocity texture from post-process
Managed to write the velocity texture to a separate texture and then
grab that pass from the post-processing shader. Right now the program
renders in three passes. First the velocity map, then the regular
shading, and finally in the post-processing shader for the effects. The
last shader is able to retrieve object data even though they are run in
separate shaders!