Applied basic motion blur filter
The applied motion filter is the bare bones of motion blur effect. Right
now the sampling follows the no dilate method which doesn't produce the
best result for motion blur. Right now a lot of tweaking is needed to
get the proper blur factor and blur intensity but the current settings
look pretty ok. Right now the filter does not care for the depth buffer
which means that objects that are in front of motion blured objects will
also be blurred.