Fixed some issues
Fixed some issues and found an ok balance between blur factor and
intensity. There are still some issues with how much blurring occurs but
it looks ok when moving in high speeds. Important to note that when
scaling objects they have to uniformarly scaled in order for the script
to work.