OMG redesigned how velocity buffers are applied
This solves so many problems that I had before. The GrabPass solution
for objects messed up how the depth buffer was collected using
replacement shaders so it didn't work in forward rendering and when I
activated deferred rendering it messed up the velocity buffer. This
solution is much more portable and manages to separate the velocity
shader from the object's material.