Added motion particles and exclude layer
Added a particle system depending on velocity, this velocity is set by
the speed effects public velocity variable. In the future this might
want to be set by the camera automatically. I've also enabled so that
the motion blur can exclude object that are in the fx_layer (like the
particles I just made).