A LOT happened... Created the demo track
Ok, so first of all I fixed an issue with the velocity shader so now it
can render as a deferred renderer as well. Still have issues where
objects have to be uniformaly big. Secondly I've created a test track
for the various aspects we want to demo on wednesday. I included a LOT
of resources for this that I found in the assets store.